DB basic CRUD

+++ app/backend/src/ban/ban.service.ts
@@ -1,8 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Admin } from 'src/entities/admin';
 import { Ban } from 'src/entities/ban';
 import { Chat } from 'src/entities/chat';
+import { ChatUsers } from 'src/entities/chat-users';
 import { Users } from 'src/entities/users';
+import { err0, err1, err2, err3, err4, err5 } from 'src/err';
 import { Repository } from 'typeorm';
 
 @Injectable()
@@ -11,48 +14,48 @@ export class BanService {
     @InjectRepository(Ban) private banRepo: Repository<Ban>,
     @InjectRepository(Users) private usersRepo: Repository<Users>,
     @InjectRepository(Chat) private chatRepo: Repository<Chat>,
+    @InjectRepository(ChatUsers) private chatUsersRepo: Repository<ChatUsers>,
+    @InjectRepository(Admin) private adminRepo: Repository<Admin>,
+
     ){}
 
   async createBan(user_id: string, channel_id: number){
     if (await this.banRepo.count({user_id: user_id, channel_id: channel_id}))  // 이미 채널의 ban 이라면
-      return false;
+      return err1;
     if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저 라면
-      return false;
+      return err2;
     if (await this.chatRepo.count({channel_id: channel_id}) === 0)  // 존재하지 않은 채널 이라면
-      return false;
+      return err4;
+    if (await this.chatUsersRepo.count({user_id: user_id, channel_id: channel_id}) === 0)  // 해당 채널에 유저가 없다면
+      return err3;
+    if (await this.adminRepo.count({user_id: user_id, channel_id: channel_id}))  // 해당 유저가 admin 이면
+      return err5;
     await this.banRepo.save({user_id: user_id, channel_id: channel_id})
-    return true;
-  }
-
-  async readBan(channel_id: number){
-    const channel = await this.banRepo.find({channel_id: channel_id});  // 해당 채널 검색
-    let bans: Array<string> = [];
-    for(var i in channel)
-      bans[i] = channel[i].user_id;
-    return bans;
+    return err0;
   }
 
   async isBan(user_id: string, channel_id: number){
     if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저 라면
+      return err2;
+    if (await this.chatRepo.count({channel_id: channel_id}) === 0)  // 존재하지 않은 채널 이라면
+      return err4;
+    if (await this.banRepo.count({user_id: user_id, channel_id: channel_id}) === 0)  // ban 당하지 않았 으면
       return false;
-    if (await this.banRepo.count({user_id: user_id, channel_id: channel_id}))
-      return true;
+    const ban = await this.banRepo.find({user_id: user_id, channel_id: channel_id});
+    const banTime = ban[0].createdAt;
+    banTime.setSeconds(banTime.getSeconds() + 60);
+    if (banTime < new Date()){  // ban 당한지 60초가 넘었으면
+      await this.banRepo.delete({user_id: user_id, channel_id: channel_id});
+      return false;
+    }
     else
-      return false;
+      return true;
   }
 
-  async deleteBan(user_id: string, channel_id: number){
+  async deleteBan(user_id: string){
     if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저 라면
-      return false;
-    if (await this.banRepo.count({ user_id: user_id, channel_id: channel_id }) === 0)  // 유저가 ban이 아니면
-      return false;
-    this.banRepo.delete({ user_id: user_id, channel_id: channel_id });
-    return true;
-  }  
-  async deleteAllBan(user_id: string){
-    if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저 라면
-      return false;
-    this.banRepo.delete({ user_id: user_id});
-    return true;
+      return err2;
+    await this.banRepo.delete({user_id: user_id});
+    return err0;
   }  
 }
